address comments; move to rhconnect.go
