+++ config/certsuite_config.yml
@@ -38,8 +38,8 @@ executedBy: ""
 partnerName: ""
 collectorAppPassword: ""
 collectorAppEndpoint: "http://claims-collector.cnf-certifications.sysdeseng.com"
-connectAPIKey: ""
 connectAPIBaseURL: "https://access.redhat.com/hydra/cwe/rest/v1.0"
+connectAPIKey: ""
 connectProjectID: ""
 connectAPIProxyURL: ""
 connectAPIProxyPort: ""
